+++ src/index.js
@@ -2,7 +2,7 @@ export default {
   async fetch(request, env) {
     const url = new URL(request.url);
 
-    if (/^\/room\/(1|2|3|4|5|6|8|9|10|11)\/?$/.test(url.pathname)) {
+    if (/^\/room\/(1|2|3|4|5|6|7|8|9|10|11)\/?$/.test(url.pathname)) {
       const roomPageUrl = new URL("/room.html", url);
       const assetRequest = new Request(roomPageUrl.toString(), {
         method: "GET",